test(hooks): consolidate symlink bin regressions in one file
Run a single incremental tsc -b before all npm-bin symlink subprocess tests to avoid parallel dist write races across Vitest workers (Greptile follow-up).
