+++ packages/cli/src/hook-bin.test.ts
@@ -1,30 +1,8 @@
-import { execFileSync, spawnSync } from "node:child_process";
-import { mkdtempSync, readFileSync, rmSync, symlinkSync } from "node:fs";
-import { tmpdir } from "node:os";
-import { dirname, join } from "node:path";
-import { fileURLToPath } from "node:url";
-import { afterEach, beforeAll, describe, expect, it } from "vitest";
+import { readFileSync } from "node:fs";
+import { describe, expect, it } from "vitest";
 import { main } from "./hook-bin.js";
 
-const itSymlink = it.skipIf(process.platform === "win32");
-
-function ensureCliDistBuilt(): void {
-  const repoRoot = join(dirname(fileURLToPath(import.meta.url)), "../../..");
-  const tscBin = join(repoRoot, "node_modules/typescript/bin/tsc");
-  execFileSync(process.execPath, [tscBin, "-b", "packages/cli"], {
-    cwd: repoRoot,
-    stdio: "pipe",
-  });
-}
-
-const temps: string[] = [];
-afterEach(() => {
-  for (const dir of temps.splice(0)) rmSync(dir, { recursive: true, force: true });
-});
-
 describe("deft-hook executable", () => {
-  const hookBinPath = join(dirname(fileURLToPath(import.meta.url)), "../dist/hook-bin.js");
-
   it("uses the direct hook dispatcher instead of the general CLI router", () => {
     const source = readFileSync(new URL("./hook-bin.ts", import.meta.url), "utf8");
     const manifest = JSON.parse(
@@ -39,59 +17,5 @@ describe("deft-hook executable", () => {
 
   it("runs the hook dispatcher without entering the general router", () => {
     expect(main([])).toBe(2);
-  });
-
-  describe("npm bin symlink entrypoint (#2846)", () => {
-    beforeAll(() => {
-      ensureCliDistBuilt();
-    });
-
-    itSymlink(
-      "runs main through an npm-style bin symlink and emits Cursor allow JSON (#2846)",
-      () => {
-        const dir = mkdtempSync(join(tmpdir(), "deft-hook-symlink-"));
-        temps.push(dir);
-        const linkPath = join(dir, "deft-hook");
-        symlinkSync(hookBinPath, linkPath);
-
-        const payload = JSON.stringify({ tool_name: "Read", workspace_root: "/project" });
-        const result = spawnSync(
-          process.execPath,
-          [linkPath, "--host=cursor", "--event=tool.before"],
-          {
-            input: payload,
-            encoding: "utf8",
-          },
-        );
-
-        expect(result.status).toBe(0);
-        expect(result.stdout.trim().length).toBeGreaterThan(0);
-        expect(JSON.parse(result.stdout)).toEqual({ permission: "allow" });
-      },
-    );
-
-    itSymlink(
-      "denies through an npm-style bin symlink instead of silent empty stdout (#2846)",
-      () => {
-        const dir = mkdtempSync(join(tmpdir(), "deft-hook-symlink-"));
-        temps.push(dir);
-        const linkPath = join(dir, "deft-hook");
-        symlinkSync(hookBinPath, linkPath);
-
-        const result = spawnSync(
-          process.execPath,
-          [linkPath, "--host=cursor", "--event=tool.before"],
-          {
-            input: "",
-            encoding: "utf8",
-          },
-        );
-
-        expect(result.status).toBe(0);
-        const decision = JSON.parse(result.stdout);
-        expect(decision.permission).toBe("deny");
-        expect(decision.user_message).toContain("stdin was empty");
-      },
-    );
   });
 });
